This commit is a checkpoint for writing backups as tar files to S3 glacier in multiple threads.
This code "works" in the sense that it can write a data to a tar archive in multiple threads to S3 using multipart uploads. However, it does not handle edges cases, is not optimized, hogs memory, has a major code smell, and contains dead code. It does provide the general idea for writing a backup as tar archives to s3 in parallel.

There is a couple current bugs, to name a few:
Files larger than 5GB throw an error due to multipart upload size limits.
Extra zeros get added which causes tar readers to think the archive is corrupted.
The buffer for the last s3 part errors if it grows larger than 5GB